ref(back): refactor folders and files structure for user router, improving organization, imports, and aligning with FastAPI async best practices
