+++ backend/src/schema/messageSchema.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Message(BaseModel):
+    message: str
