+++ backend/src/schema/userSchema.py
@@ -1,27 +1,34 @@
-# from pydantic import BaseModel, EmailStr
+from pydantic import BaseModel, ConfigDict, EmailStr
 
 
-# class Message(BaseModel):
-#     message: str
+class Message(BaseModel):
+    message: str
 
 
-# class UserSchema(BaseModel):
-#     user_name: str
-#     user_nickname: str
-#     user_email: EmailStr
-#     user_password: str
+class UserSchema(BaseModel):
+    user_name: str
+    user_nickname: str
+    user_email: EmailStr
+    user_password: str
 
 
-# class UserPublic(BaseModel):
-#     user_id: int
-#     user_name: str
-#     user_nickname: str
-#     user_email: EmailStr
+class UserPublic(BaseModel):
+    user_id: int
+    user_name: str
+    user_nickname: str
+    user_email: EmailStr
+    model_config = ConfigDict(from_attributes=True)
 
 
-# class UserDB(UserSchema):
-#     user_id: int
+class UserList(BaseModel):
+    users: list[UserPublic]
 
 
-# class UserList(BaseModel):
-#     users: list[UserPublic]
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class FilterPage(BaseModel):
+    offset: int = 0
+    limit: int = 100
