--- backend/src/schemas.py
@@ -1,34 +1,0 @@
-from pydantic import BaseModel, ConfigDict, EmailStr
-
-
-class Message(BaseModel):
-    message: str
-
-
-class UserSchema(BaseModel):
-    user_name: str
-    user_nickname: str
-    user_email: EmailStr
-    user_password: str
-
-
-class UserPublic(BaseModel):
-    user_id: int
-    user_name: str
-    user_nickname: str
-    user_email: EmailStr
-    model_config = ConfigDict(from_attributes=True)
-
-
-class UserList(BaseModel):
-    users: list[UserPublic]
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class FilterPage(BaseModel):
-    offset: int = 0
-    limit: int = 100
