+++ backend/tests/conftest.py
@@ -11,7 +11,7 @@ from sqlalchemy.pool import StaticPool
 
 from src.app import app
 from src.database import get_session
-from src.models import User, table_registry
+from src.model.user import User, table_registry
 from src.security import get_password_hash
